Adapt documentation from "QuantumCitations" fork (#4)
This makes the documentation appropriate for QuantumCitations having
been merged back into DocumenterCitations.

The default style (following QuantumCitations) has changed to
`:numeric`, making this a breaking change. The differences to the
previous version of DocumenterCitations are outlined in `NEWS.md`,
serving as a guideline for any users switching to the next release of
DocumenterCitations, but wanting to keep the existing author-year style
citations.