Prevent duplicate labels in `:alpha` style
This implements a stateful `AlphaStyle` that disambiguates the labels
for the bibliography by appending a suffix "a", "b", etc. This emulates
the behavior of the numeric citation style in LaTeX.

To enable the generating of unique labels, there is a new internal
`init_bibliography!` function that is called at the beginning of the
`ExpandBibliography` pipeline step.

The existing `:alpha` style is automatically upgraded to the new
`AlphaStyle`. It's just easier for users to pass `style=:alpha` than to
instantiate the full `AlphaStyle`. Moreover, there's no reason anyone
would ever want to use the old "dumb" `:alpha` style.

In addition to the new `AlphaStyle`, this also fixes some
inconsistencies relative to the alphabetic citation style in LaTeX:

* Up to four names are included in the label, not 3 (but 5 or more names
  results in 3 names and "+")
* The first letter of a "particle" of the name is included in the label,
  e.g. "vWB08" for a first author "von Winckel"