Improve parsing of tex strings
Implements a proper (albeit limited) parser for the subset of tex syntax
that occurs in BibTeX fields, converting to the equivalent Markdown.
Most importantly, this handles accents like \"o, and groupings, i.e., it
removes curly braces in the same way that LaTeX does.

Inline math is recognized and translated into Documenter's preferred
"``" syntax, but otherwise left alone.

A limited set of LaTeX commands is recognized, most notably \href and
\url, allowing to put hyperlinks in the `@note` field of a reference.
We're deliberately conservative here. Anyone who can demonstrate the
need for additional commands should submit an issue with a real-life
example. It should be quite trivial to add any specific additional
command.

All of this is considered "very internal". We're not making any
promises. Ideally, a lot of this code would probably be better in
Bibliography.jl, but at this point, we're treating Bibliography as a
just a basic parser to get the raw tex, and doing all the processing /
conversion to Markdown ourselves.