Remove workarounds for bugs fixed in BibInternal 0.3.5
The release of `BibInternal v0.3.5` solved issue
https://github.com/Humans-of-Julia/BibInternal.jl/issues/22 related to
preprints. Our monkey-patch of `BibInternal.make_bibtex_entry` to work
around this is no longer necessary.

The release also solved issue
https://github.com/Humans-of-Julia/BibInternal.jl/pull/21, so the
description of the workaround in the documentation can be removed.

Note that `BibInternal` is only an implicit dependency (via
`Bibliography.jl`). The lowest compat-bound for `Bibliography.jl`,
that is, `v0.2.15`, is compatible with `BibInternal v0.3.5`. So that
version should be installed automatically. Someone would have to go out
of their way to pin an older version of `BibInternal` in their
environment. Dealing with any resulting bugs in that case is on them.