Fix rendering of non-article references.
This brings the rendering of non-article entries in line with RevTeX. It
significantly extends the support especially for `@inproceedings`,
`@incollection`, `@inbook`. The `Chapter`, `Volume`, `Number`,
`Edition`, `Month`, fields are now supported for several entry types.

Revises how the URL/DOI is linked in the references.

Adds the `title_transform_case` option to the internal formatting
routines, for use by custom styles. In principle, this could be used to
implement the sentence-case transformation that is common in BibTeX
styles. Note that I don't condone this convention: The titles should
appear as they do in the published material (which generally means title
case, not sentence case). In any case, `DocumenterCitations` will
continue to render titles as they appear in the `.bib` file, making
braces to "protect" capitalized words unnecessary.