Show error file paths consistently with Documenter
In v1.10, Documenter changed paths in error messages from being relative
to `make.jl` to being relative to the CWD, via the internal
`Documenter.locrepr` function. This PR consistenly uses
`Documenter.locrepr` for file paths in error messages, ensuring that
`DocumenterCitations` is consistent with `Documenter` both pre-1.10 and
post-1.10.

As a side effect, the location of a `@bibliography` block containing
errors is now shown including line numbers in more situations.